requirements-novo.txt
Flask==3.1.1
python-dotenv==1.1.1
weasyprint==66.0
docx2pdf==0.1.8
requests==2.32.4
lxml==6.0.1
cffi==1.17.1
pydyf==0.11.0
tinycss2==1.4.0
cssselect2==0.8.0
Pillow==11.3.0
fonttools==4.59.2
html5lib==1.1
pyphen==0.17.2
tinyhtml5==2.0.0
weblink==0.1.0
# LibreOffice needed for DOCX conversion on Linux
